# Conflicts:
#	app/src/main/java/com/cylan/jiafeigou/cache/db/module/tasks/DPCamMultiQueryTask.java
#	app/src/main/java/com/cylan/jiafeigou/cache/db/module/tasks/DPSingleQueryTask.java
#	app/src/main/java/com/cylan/jiafeigou/dp/DataPoint.java
#	app/src/main/java/com/cylan/jiafeigou/n/engine/DataSourceService.java
#	app/src/main/java/com/cylan/jiafeigou/n/mvp/impl/cam/CamLivePresenterImpl.java
#	app/src/main/java/com/cylan/jiafeigou/n/mvp/impl/cam/CamMessageListPresenterImpl.java
#	app/src/main/java/com/cylan/jiafeigou/n/mvp/impl/cam/CamSettingPresenterImpl.java
#	app/src/main/java/com/cylan/jiafeigou/n/mvp/impl/setting/SafeInfoPresenterImpl.java
#	app/src/main/java/com/cylan/jiafeigou/n/view/adapter/CamMessageListAdapter.java
#	app/src/main/java/com/cylan/jiafeigou/n/view/cam/CamLiveController.java
#	app/src/main/java/com/cylan/jiafeigou/n/view/cam/CapturePeriodDialogFragment.java
#	app/src/main/java/com/cylan/jiafeigou/n/view/cam/DeviceInfoDetailFragment.java
#	app/src/main/java/com/cylan/jiafeigou/n/view/cam/DeviceTimeZoneFragment.java
#	app/src/main/java/com/cylan/jiafeigou/n/view/cam/SafeProtectionFragment.java
#	app/src/main/java/com/cylan/jiafeigou/utils/MiscUtils.java
#	app/src/main/java/com/cylan/jiafeigou/widget/dialog/TimePickDialogFragment.java

+++ app/src/main/java/com/cylan/jiafeigou/cache/db/module/tasks/BaseDPTask.java
@@ -61,9 +61,9 @@ public abstract class BaseDPTask<T extends IDPTaskResult> implements IDPSingleTa
                 .filter(rsp -> rsp.seq == seq).first().timeout(GLOBAL_NET_OPERATION_TIME_OUT, TimeUnit.SECONDS);
     }
 
-    protected Observable<RxEvent.SetDataRsp> makeSetDataRspResponse(int seq) {
+    protected Observable<RxEvent.SetDataRsp> makeSetDataRspResponse(long seq) {
         return RxBus.getCacheInstance().toObservable(RxEvent.SetDataRsp.class)
-                .filter(rsp -> ((int) rsp.seq) == seq).first().timeout(GLOBAL_NET_OPERATION_TIME_OUT, TimeUnit.SECONDS);
+                .filter(rsp -> (rsp.seq) == seq).first().timeout(GLOBAL_NET_OPERATION_TIME_OUT, TimeUnit.SECONDS);
     }
 
     protected Observable makeGetDataRequest() {
